Refactor the architecture of scylladb codebase
Separate codes by their functionality.
ta_storage encapsulates Scylladb oriented headers
scylladb_client is for connection Scylladb host.
scylladb_identity is for reattachment.
scylladb_permanode is for Permanode.
scylladb_utils is for common functions and definitions.

close #365
