+++ storage/scylladb_permanode.c
@@ -5,23 +5,10 @@
  * terms of the MIT license. A copy of the license can be found in the file
  * "LICENSE" at the root of this distribution.
  */
-#include "scylla_api.h"
-#include <time.h>
 
-#define SCYLLA_API_LOGGER "scylladb"
+#include "scylladb_permanode.h"
 
-static logger_id_t logger_id;
-
-void scylla_api_logger_init() { logger_id = logger_helper_enable(SCYLLA_API_LOGGER, LOGGER_DEBUG, true); }
-
-int scylla_api_logger_release() {
-  logger_helper_release(logger_id);
-  if (logger_helper_destroy() != RC_OK) {
-    log_critical(logger_id, "%s.\n", SCYLLA_API_LOGGER);
-    return EXIT_FAILURE;
-  }
-  return 0;
-}
+#define logger_id scylladb_logger_id
 
 typedef struct select_where_s {
   cass_byte_t* bundle;
@@ -255,128 +242,12 @@ static void print_error(CassFuture* future) {
   ta_log_error("Error: %.*s\n", (int)message_length, message);
 }
 
-static CassCluster* create_cluster(const char* hosts) {
-  CassCluster* cluster = cass_cluster_new();
-  cass_cluster_set_contact_points(cluster, hosts);
-  return cluster;
-}
-
-static CassError connect_session(CassSession* session, const CassCluster* cluster) {
-  CassError rc = CASS_OK;
-  CassFuture* future = cass_session_connect(session, cluster);
-
-  cass_future_wait(future);
-  rc = cass_future_error_code(future);
-  if (rc != CASS_OK) {
-    print_error(future);
-  }
-  cass_future_free(future);
-
-  return rc;
-}
-
-static CassError execute_query(CassSession* session, const char* query) {
-  CassError rc = CASS_OK;
-  CassStatement* statement = cass_statement_new(query, 0);
-  CassFuture* future = cass_session_execute(session, statement);
-
-  cass_future_wait(future);
-
-  rc = cass_future_error_code(future);
-  if (rc != CASS_OK) {
-    print_error(future);
-  }
-
-  cass_future_free(future);
-  cass_statement_free(statement);
-
-  return rc;
-}
-
-static CassError prepare_query(CassSession* session, const char* query, const CassPrepared** prepared) {
-  CassError rc = CASS_OK;
-  CassFuture* future = NULL;
-
-  future = cass_session_prepare(session, query);
-  cass_future_wait(future);
-
-  rc = cass_future_error_code(future);
-  if (rc != CASS_OK) {
-    print_error(future);
-  } else {
-    *prepared = cass_future_get_prepared(future);
-  }
-
-  cass_future_free(future);
-
-  return rc;
-}
-
-static CassError execute_statement(CassSession* session, CassStatement* statement) {
-  CassError rc = CASS_OK;
-  CassFuture* future = NULL;
-  future = cass_session_execute(session, statement);
-
-  cass_future_wait(future);
-
-  rc = cass_future_error_code(future);
-  if (rc != CASS_OK) {
-    print_error(future);
-  }
-
-  cass_future_free(future);
-  cass_statement_free(statement);
-
-  return rc;
-}
-
 static status_t get_blob(const CassRow* row, cass_byte_t* target, const char* column) {
   size_t len;
   const cass_byte_t* buf;
   status_t ret = SC_OK;
   cass_value_get_bytes(cass_row_get_column_by_name(row, column), &buf, &len);
   memcpy(target, buf, len);
-  return ret;
-}
-
-static status_t make_query(char** result, const char* head_desc, const char* position, const char* left_desc) {
-  if (head_desc == NULL || position == NULL || left_desc == NULL) {
-    ta_log_error("NULL pointer to CQL query\n");
-    return SC_TA_NULL;
-  }
-  size_t head_len = strlen(head_desc);
-  size_t pos_len = strlen(position);
-  size_t left_len = strlen(left_desc);
-  size_t result_len = head_len + pos_len + left_len + 1;
-  *result = malloc(result_len * sizeof(char));
-  if (*result == NULL) {
-    ta_log_error("%s\n", "SC_STORAGE_OOM");
-    return SC_STORAGE_OOM;
-  }
-  memcpy(*result, head_desc, head_len);
-  memcpy(*result + head_len, position, pos_len);
-  memcpy(*result + head_len + pos_len, left_desc, left_len);
-  (*result)[result_len - 1] = 0;
-
-  return SC_OK;
-}
-
-static status_t create_keyspace(CassSession* session, const char* keyspace_name) {
-  status_t ret = SC_OK;
-  char* create_query = NULL;
-  ret = make_query(&create_query, "CREATE KEYSPACE IF NOT EXISTS ", keyspace_name,
-                   " WITH replication = {"
-                   "'class': 'SimpleStrategy', 'replication_factor': '2'};");
-  if (ret != SC_OK) {
-    ta_log_error("%s\n", "make Create keyspace query fail");
-    return ret;
-  }
-  if (execute_query(session, create_query) != CASS_OK) {
-    ta_log_error("Create keyspace %s fail\n", keyspace_name);
-    ret = SC_STORAGE_CASSANDRA_QUREY_FAIL;
-  }
-
-  free(create_query);
   return ret;
 }
 
@@ -433,136 +304,6 @@ exit:
   return ret;
 }
 
-static status_t create_identity_table(CassSession* session, bool need_drop) {
-  if (need_drop) {
-    if (execute_query(session, "DROP TABLE IF EXISTS identity;") != CASS_OK) {
-      ta_log_error("drop identity table fail\n");
-      return SC_STORAGE_CASSANDRA_QUREY_FAIL;
-    }
-  }
-  if (execute_query(session,
-                    "CREATE TABLE IF NOT EXISTS identity("
-                    "id bigint, ts timestamp, hash blob, status tinyint, PRIMARY KEY (id));") != CASS_OK) {
-    ta_log_error("create identity table fail\n");
-    return SC_STORAGE_CASSANDRA_QUREY_FAIL;
-  }
-  if (execute_query(session, "CREATE INDEX IF NOT EXISTS ON identity(status);") != CASS_OK) {
-    ta_log_error("create identity table index fail\n");
-    return SC_STORAGE_CASSANDRA_QUREY_FAIL;
-  }
-  if (execute_query(session, "CREATE INDEX IF NOT EXISTS ON identity(hash);") != CASS_OK) {
-    ta_log_error("create identity table index fail\n");
-    return SC_STORAGE_CASSANDRA_QUREY_FAIL;
-  }
-
-  return SC_OK;
-}
-
-status_t db_client_service_init(db_client_service_t* service) {
-  if (service == NULL) {
-    ta_log_error("NULL pointer to ScyllaDB client service for connection endpoint(s)\n");
-    return SC_TA_NULL;
-  }
-  if (service->host == NULL) {
-    ta_log_error("NULL pointer to ScyllaDB hostname\n");
-    return SC_TA_NULL;
-  }
-  service->session = cass_session_new();
-  service->cluster = create_cluster(service->host);
-  if (connect_session(service->session, service->cluster) != CASS_OK) {
-    ta_log_error("connect ScyllaDB cluster with host : %s failed\n", service->host);
-    service->enabled = false;
-    return SC_STORAGE_CONNECT_FAIL;
-  }
-  service->enabled = true;
-  return SC_OK;
-}
-
-status_t db_client_service_free(db_client_service_t* service) {
-  if (service == NULL) {
-    ta_log_error("NULL pointer to ScyllaDB client service for connection endpoint(s)\n");
-    return SC_TA_NULL;
-  }
-  if (service->session != NULL) {
-    cass_session_free(service->session);
-  }
-  if (service->cluster != NULL) {
-    cass_cluster_free(service->cluster);
-  }
-  free(service->host);
-  return SC_OK;
-}
-
-status_t db_insert_tx_into_identity(db_client_service_t* service, int64_t id, const char* hash,
-                                    db_txn_status_t status) {
-  status_t ret = SC_OK;
-  db_identity_t* identity = NULL;
-
-  if ((ret = db_identity_new(&identity)) != SC_OK) {
-    ta_log_error("db new identity failed\n");
-    goto exit;
-  }
-  if ((ret = db_set_identity_hash(identity, (cass_byte_t*)hash, NUM_FLEX_TRITS_HASH)) != SC_OK) {
-    ta_log_error("db set identity hash failed\n");
-    goto exit;
-  }
-  if ((ret = db_set_identity_id(identity, id)) != SC_OK) {
-    ta_log_error("db set identity id failed\n");
-    goto exit;
-  }
-  if ((ret = db_set_identity_status(identity, status)) != SC_OK) {
-    ta_log_error("db set identity status failed\n");
-    goto exit;
-  }
-
-  if ((ret = db_set_identity_timestamp(identity, (cass_int64_t)time(NULL))) != SC_OK) {
-    ta_log_error("db set identity time failed\n");
-    goto exit;
-  }
-  if ((ret = db_insert_identity_table(service, identity)) != SC_OK) {
-    ta_log_error("db insert identity table failed\n");
-    goto exit;
-  }
-
-exit:
-  db_identity_free(&identity);
-  return ret;
-}
-
-status_t db_init_identity_keyspace(db_client_service_t* service, bool need_drop, const char* keyspace_name) {
-  status_t ret = SC_OK;
-  CassStatement* use_statement = NULL;
-  char* use_query = NULL;
-  if (service == NULL) {
-    ta_log_error("NULL pointer to ScyllaDB client service for connection endpoint(s)\n");
-    return SC_TA_NULL;
-  }
-  if ((ret = create_keyspace(service->session, keyspace_name)) != SC_OK) {
-    ta_log_error("create %s keyspace fail\n", keyspace_name);
-    goto exit;
-  }
-  ret = make_query(&use_query, "USE ", keyspace_name, "");
-  if (ret != SC_OK) {
-    ta_log_error("%s\n", "make USE keyspace query fail");
-    goto exit;
-  }
-  use_statement = cass_statement_new(use_query, 0);
-  if (execute_statement(service->session, use_statement) != CASS_OK) {
-    ta_log_error("USE keyspace %s fail\n", keyspace_name);
-    ret = SC_STORAGE_CASSANDRA_QUREY_FAIL;
-    goto exit;
-  }
-
-  if ((ret = create_identity_table(service->session, need_drop)) != SC_OK) {
-    ta_log_error("%s\n", "create identity table fail");
-    goto exit;
-  }
-
-exit:
-  free(use_query);
-  return ret;
-}
-
 status_t db_permanent_keyspace_init(db_client_service_t* service, bool need_drop, const char* keyspace_name) {
   status_t ret = SC_OK;
   CassStatement* use_statement = NULL;
@@ -599,158 +340,6 @@ status_t db_permanent_keyspace_init(db_client_service_t* service, bool need_drop
 
 exit:
   free(use_query);
-  return ret;
-}
-
-static CassStatement* ret_insert_identity_statement(const CassPrepared* prepared, const db_identity_t* obj) {
-  CassStatement* statement = NULL;
-  statement = cass_prepared_bind(prepared);
-  cass_statement_bind_int64_by_name(statement, "id", db_ret_identity_id(obj));
-  cass_statement_bind_int64_by_name(statement, "ts", db_ret_identity_timestamp(obj));
-  cass_statement_bind_bytes_by_name(statement, "hash", db_ret_identity_hash(obj), NUM_FLEX_TRITS_HASH);
-  cass_statement_bind_int8_by_name(statement, "status", db_ret_identity_status(obj));
-  return statement;
-}
-
-status_t db_insert_identity_table(db_client_service_t* service, db_identity_t* obj) {
-  status_t ret = SC_OK;
-  const CassPrepared* insert_prepared = NULL;
-  CassStatement* statement = NULL;
-  const char* insert_query =
-      "INSERT INTO identity (id, ts, status, hash)"
-      "VALUES (?, ?, ?, ?);";
-  if (service == NULL) {
-    ta_log_error("NULL pointer to ScyllaDB client service for connection endpoint(s)");
-    return SC_TA_NULL;
-  }
-  if (obj == NULL) {
-    ta_log_error("NULL pointer to ScyllaDB identity object\n");
-    return SC_TA_NULL;
-  }
-
-  if (prepare_query(service->session, insert_query, &insert_prepared) != CASS_OK) {
-    ta_log_error("%s\n", "prepare INSERT query fail");
-    return SC_STORAGE_CASSANDRA_QUREY_FAIL;
-  }
-  statement = ret_insert_identity_statement(insert_prepared, obj);
-  if (execute_statement(service->session, statement) != CASS_OK) {
-    ta_log_error("insert obj into identity table fail\n");
-    ret = SC_STORAGE_CASSANDRA_QUREY_FAIL;
-    goto exit;
-  }
-
-exit:
-  cass_prepared_free(insert_prepared);
-  return ret;
-}
-
-static status_t get_identity_array(CassSession* session, CassStatement* statement,
-                                   db_identity_array_t* identity_array) {
-  status_t ret = SC_OK;
-  CassFuture* future = NULL;
-  const CassResult* result;
-  CassIterator* iterator;
-  db_identity_t* identity = NULL;
-  if ((ret = db_identity_new(&identity)) != SC_OK) {
-    goto exit;
-  }
-  future = cass_session_execute(session, statement);
-  cass_future_wait(future);
-  if (cass_future_error_code(future) != CASS_OK) {
-    print_error(future);
-    ret = SC_STORAGE_CASSANDRA_QUREY_FAIL;
-    goto exit;
-  }
-
-  result = cass_future_get_result(future);
-  iterator = cass_iterator_from_result(result);
-
-  while (cass_iterator_next(iterator)) {
-    db_identity_t* itr;
-    const cass_byte_t* hash;
-    cass_int8_t value;
-    cass_int64_t id;
-    cass_int64_t ts;
-    size_t len;
-    const CassRow* row = cass_iterator_get_row(iterator);
-
-    utarray_extend_back(identity_array);
-    itr = (db_identity_t*)utarray_back(identity_array);
-
-    cass_value_get_bytes(cass_row_get_column_by_name(row, "hash"), &hash, &len);
-    db_set_identity_hash(itr, hash, NUM_FLEX_TRITS_HASH);
-    cass_value_get_int8(cass_row_get_column_by_name(row, "status"), &value);
-    db_set_identity_status(itr, value);
-    cass_value_get_int64(cass_row_get_column_by_name(row, "id"), &id);
-    db_set_identity_id(itr, id);
-    cass_value_get_int64(cass_row_get_column_by_name(row, "ts"), &ts);
-    db_set_identity_timestamp(itr, ts);
-  }
-
-  cass_result_free(result);
-  cass_iterator_free(iterator);
-
-exit:
-  cass_future_free(future);
-  cass_statement_free(statement);
-  db_identity_free(&identity);
-  return ret;
-}
-
-status_t db_get_identity_objs_by_status(db_client_service_t* service, cass_int8_t status,
-                                        db_identity_array_t* identity_array) {
-  status_t ret = SC_OK;
-  CassStatement* statement = NULL;
-  const CassPrepared* select_prepared = NULL;
-  const char* query = "SELECT * FROM identity WHERE status = ?;";
-
-  if (prepare_query(service->session, query, &select_prepared) != CASS_OK) {
-    ta_log_error("%s\n", "prepare SELECT query fail");
-    return SC_STORAGE_CASSANDRA_QUREY_FAIL;
-  }
-  statement = cass_prepared_bind(select_prepared);
-  cass_statement_bind_int8_by_name(statement, "status", status);
-  get_identity_array(service->session, statement, identity_array);
-
-  cass_prepared_free(select_prepared);
-  return ret;
-}
-
-status_t db_get_identity_objs_by_id(db_client_service_t* service, cass_int64_t id,
-                                    db_identity_array_t* identity_array) {
-  status_t ret = SC_OK;
-  CassStatement* statement = NULL;
-  const CassPrepared* select_prepared = NULL;
-  const char* query = "SELECT * FROM identity WHERE id = ?;";
-
-  if (prepare_query(service->session, query, &select_prepared) != CASS_OK) {
-    ta_log_error("%s\n", "prepare SELECT query fail");
-    return SC_STORAGE_CASSANDRA_QUREY_FAIL;
-  }
-  statement = cass_prepared_bind(select_prepared);
-  cass_statement_bind_int64_by_name(statement, "id", id);
-  get_identity_array(service->session, statement, identity_array);
-
-  cass_prepared_free(select_prepared);
-  return ret;
-}
-
-status_t db_get_identity_objs_by_hash(db_client_service_t* service, const cass_byte_t* hash,
-                                      db_identity_array_t* identity_array) {
-  status_t ret = SC_OK;
-  CassStatement* statement = NULL;
-  const CassPrepared* select_prepared = NULL;
-  const char* query = "SELECT * FROM identity WHERE hash = ?;";
-
-  if (prepare_query(service->session, query, &select_prepared) != CASS_OK) {
-    ta_log_error("%s\n", "prepare SELECT query fail");
-    return SC_STORAGE_CASSANDRA_QUREY_FAIL;
-  }
-  statement = cass_prepared_bind(select_prepared);
-  cass_statement_bind_bytes_by_name(statement, "hash", hash, NUM_FLEX_TRITS_HASH);
-  get_identity_array(service->session, statement, identity_array);
-
-  cass_prepared_free(select_prepared);
   return ret;
 }
