+++ storage/ta_storage.h
@@ -1,0 +1,22 @@
+/*
+ * Copyright (C) 2019 BiiLabs Co., Ltd. and Contributors
+ * All Rights Reserved.
+ * This is free software; you can redistribute it and/or modify it under the
+ * terms of the MIT license. A copy of the license can be found in the file
+ * "LICENSE" at the root of this distribution.
+ */
+#ifndef TA_STORAGE_H_
+#define TA_SCYLLADB_TA_API_H_
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include "scylladb_client.h"
+#include "scylladb_identity.h"
+#include "scylladb_permanode.h"
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif  // TA_STORAGE_H_
